+++ src/pages/member/Login.scss
@@ -2,24 +2,4 @@
     display: flex;
     justify-content: center;
     padding-top: 100px;
-
-    .title {
-        margin-bottom: 36px;
-        line-height: 1em;
-        color: rgb(0, 0, 0);
-        font-size: 26px;
-        font-weight: bold;
-        text-align: center;
-    }
-
-    .otherBtn {
-        width: 460px;
-        height: 48px;
-        margin-top: 15px;
-        background-color: rgb(248, 248, 249);
-        color: rgb(0, 0, 0);
-        font-weight: bold;
-        cursor: pointer;
-        user-select: none;
-    }
 }
