+++ src/pages/member/LoginTitle.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import './LoginTitle.scss';
+
+class LoginTitle extends Component {
+  render() {
+    return (
+      <div className="LoginTitle">
+        <header className="title">로그인</header>
+      </div>
+    );
+  }
+}
+
+export default LoginTitle;
