+++ src/pages/member/LoginTitle.scss
@@ -1,0 +1,10 @@
+.LoginTitle {
+    .title {
+        margin-bottom: 36px;
+        line-height: 1em;
+        color: rgb(0, 0, 0);
+        font-size: 26px;
+        font-weight: bold;
+        text-align: center;
+    }
+}
