--- src/pages/member/Member.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Member extends Component {
-  render() {
-    return (
-      <div>
-        <p>멤버 페이지</p>
-      </div>
-    );
-  }
-}
-
-export default Member;
