+++ src/pages/member/OtherLoginButton.scss
@@ -1,0 +1,12 @@
+.OtherLoginButton {
+    .otherLoginBtn {
+        width: 460px;
+        height: 48px;
+        margin-top: 15px;
+        background-color: rgb(248, 248, 249);
+        color: rgb(0, 0, 0);
+        font-weight: bold;
+        cursor: pointer;
+        user-select: none;
+    }
+}
